+++ src/view/BookListener.java
@@ -1,5 +1,5 @@
 package view;
 
 public interface BookListener {
-	public void formEventHandler(String title, String author, int isbn, String genre);
+	public void formEventHandler(String title, String author, long isbn, String genre);
 }
